requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-filter==25.1
Pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.0.1
django-unfold==0.8.0
drf-yasg==1.21.7 